feat: modularise webview code under src/webview/ and update all imports
- Moved webviewManager.ts to src/webview/
- Updated all related imports in the codebase to reflect the new structure
- Resolved all modularisation and import path errors
- All webview logic is now fully isolated under src/webview/ for maintainability and Cursor/VS Code compatibility
- Build, lint, and type checks are clean; only a known VS Code test runner issue remains (unrelated to migration)
- Updated migration logs in memory-bank/progress/current.md and docs/wip/RESTRUCTURE_PLAN.md
