+++ package.json
@@ -6,7 +6,7 @@
     "url": "https://github.com/sm-moshi/aimemory"
   },
   "description": "Easily manage AI context for your projects using memory bank technique",
-  "version": "0.2.1",
+  "version": "0.2.4",
   "engines": {
     "vscode": "^1.96.2"
   },
