+++ src/webview/package.json
@@ -1,7 +1,7 @@
 {
   "name": "aimemory-webview",
   "private": true,
-  "version": "0.2.0",
+  "version": "0.2.4",
   "type": "module",
   "scripts": {
     "dev": "vite",
